+++ actionpack/CHANGELOG.md
@@ -722,4 +722,6 @@
 *   `ActionView::Helpers::TextHelper#highlight` now defaults to the
     HTML5 `mark` element. *Brian Cardarella*
 
+*   Prevent raising EOFError on multipart GET request (IE issue). *Adam Stankiewicz*
+
 Please check [3-2-stable](https://github.com/rails/rails/blob/3-2-stable/actionpack/CHANGELOG.md) for previous changes.